New lua interfaces, bug fixes.
New lua interfaces, bug fixes.
Debug hud improvements
Color ramp texture bug fixes
skipping mesh rendering by distance and radius.
Alpha meterial cloning bug fix
Run time shader define control for praticle materials
Trail render object bug fixes and improvements

Force compound option for collision shapes
Transformation reflections for attached rigid bodies
Shader improvements

Added KeyValueDataSet for ListBox
Added UIProperty NO_FOCUS_BY_CLICK

+++ Code/Engine/Voxelizer.cpp
@@ -94,8 +94,8 @@ void Voxelizer::CalcDistanceMap()
 	pDepthRT->SetDepthStencilDesc(mNumVoxels, mNumVoxels, PIXEL_FORMAT_D32_FLOAT, false, false);
 	ICamera* pCam = pDepthRT->GetCamera();
 	pCam->SetOrthogonal(true);
-	pCam->SetWidth((int)(radius*2.0f));
-	pCam->SetHeight((int)(radius*2.0f));
+	pCam->SetWidth(radius*2.0f);
+	pCam->SetHeight(radius*2.0f);
 	pCam->SetNearFar(-radius, radius);
 	pCam->SetPos(Vec3(0, 0, 0));
 	auto scene = pDepthRT->CreateScene();
